chore: ultra device and com settings

+++ app/device/device_exp_ultrasound_settings.h
@@ -34,9 +34,9 @@ class DeviceUltrasound {
   int32_t exp_clipping_enable_;
 
   /// @brief  exp clipping time duration.
-  int64_t exp_clip_time_duration_;
+  int32_t exp_clip_time_duration_;
   /// @brief  exp clipping time paused.
-  int64_t exp_clip_time_paused_;
+  int32_t exp_clip_time_paused_;
 
   /// @brief  exp max cycle count.
   int64_t exp_max_cycle_count_;
